requirements.txt
scipy==1.9.3
pyyaml==6.0
fastapi==0.85.1
python-multipart==0.0.5
uvicorn[standard]
h5py==3.7.0
httptools==0.5.0
keras==2.9.0
Keras-Preprocessing==1.1.2
Pillow==9.3.0
pip==22.2.2
requests==2.28.1
requests-oauthlib==1.3.1
scikit-image==0.19.3
setuptools==63.2.0
split-folders==0.5.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.0
tensorflow-io-gcs-filesystem==0.28.0
uvicorn==0.20.0
wheel==0.38.4
